+++ Scripts/Main.gd
@@ -76,7 +76,9 @@ func _ready() -> void:
 		"Outline" : 0
 		}
 	var help_menu_items := {
+		"View Splash Screen" : 0,
 		"Issue Tracker" : 0,
+		"Changelog" : 0,
 		"About Pixelorama" : 0
 		}
 
@@ -337,9 +339,14 @@ func image_menu_id_pressed(id : int) -> void:
 
 func help_menu_id_pressed(id : int) -> void:
 	match id:
-		0: # Issue Tracker
+		0: # Splash Screen
+			$SplashDialog.popup_centered() # Splash screen
+			Global.can_draw = false
+		1: # Issue Tracker
 			OS.shell_open("https://github.com/Orama-Interactive/Pixelorama/issues")
-		1: # About Pixelorama
+		2: # Changelog
+			OS.shell_open("https://github.com/Orama-Interactive/Pixelorama/blob/master/Changelog.md")
+		3: # About Pixelorama
 			$AboutDialog.popup_centered()
 			Global.can_draw = false